#2636 break-out class into separate file
