+++ src/psyclone/domain/lfric/__init__.py
@@ -51,6 +51,7 @@ from psyclone.domain.lfric.kern_call_acc_arg_list import KernCallAccArgList
 from psyclone.domain.lfric.kern_call_invoke_arg_list import \
     KernCallInvokeArgList
 from psyclone.domain.lfric.kernel_interface import KernelInterface
+from psyclone.domain.lfric.lfric_cell_iterators import LFRicCellIterators
 from psyclone.domain.lfric.lfric_extract_driver_creator import \
     LFRicExtractDriverCreator
 from psyclone.domain.lfric.lfric_symbol_table import LFRicSymbolTable
@@ -84,6 +85,7 @@ __all__ = [
     'KernelInterface',
     'KernStubArgList',
     'LFRicArgDescriptor',
+    'LFRicCellIterators',
     'LFRicCollection',
     'LFRicConstants',
     'LFRicDofmaps',
