+++ src/psyclone/domain/lfric/lfric_kern.py
@@ -44,8 +44,10 @@ from collections import OrderedDict, namedtuple
 
 from psyclone.configuration import Config
 from psyclone.core import AccessType
-from psyclone.domain.lfric import (KernCallArgList, KernStubArgList,
-                                   KernelInterface, LFRicConstants)
+from psyclone.domain.lfric.kern_call_arg_list import KernCallArgList
+from psyclone.domain.lfric.lfric_constants import LFRicConstants
+from psyclone.domain.lfric.kern_stub_arg_list import KernStubArgList
+from psyclone.domain.lfric.kernel_interface import KernelInterface
 from psyclone.errors import GenerationError, InternalError, FieldNotFoundError
 from psyclone.f2pygen import ModuleGen, SubroutineGen, UseGen
 from psyclone.parse.algorithm import Arg, KernelCall
@@ -601,9 +603,10 @@ class LFRicKern(CodedKern):
         # Add all the declarations
         # Import here to avoid circular dependency
         # pylint: disable=import-outside-toplevel
-        from psyclone.domain.lfric import (LFRicScalarArgs, LFRicFields,
-                                           LFRicDofmaps, LFRicStencils)
-        from psyclone.dynamo0p3 import (LFRicCellIterators, DynFunctionSpaces,
+        from psyclone.domain.lfric import (
+            LFRicCellIterators, LFRicScalarArgs, LFRicFields,
+            LFRicDofmaps, LFRicStencils)
+        from psyclone.dynamo0p3 import (DynFunctionSpaces,
                                         DynCMAOperators, DynBoundaryConditions,
                                         DynLMAOperators, LFRicMeshProperties,
                                         DynBasisFunctions, DynReferenceElement)
